refactor: extract admin router (PR-3, Phase 1)
Pure move of the 5 /api/admin/* endpoints from server/routes.ts into
server/routes/admin.ts as a default-export Router with per-route
middleware (requireClerkUser + requireAdmin). Mounted via app.use at the
position of the first admin endpoint (/api/admin/health) to preserve
registration order. Handlers moved verbatim; removed the newly-unused
ActionsConfigSchema import from routes.ts.
